pull request some refactor #37
