+++ app/auth/signup/page.tsx
@@ -1,5 +1,5 @@
-import { SingUpPage } from '@/pages/auth/signUp/singUpPage'
+import { SignUpPage } from '@/pages/auth/signUp/signUpPage'
 
 export default function SignUp() {
-  return <SingUpPage />
+  return <SignUpPage />
 }
