+++ src/pages/auth/signIn/signInPage.tsx
@@ -23,7 +23,6 @@ const SignInPage = () => {
 
   const onSubmit = (data: LoginFormSchema) => {
     signIn(data)
-    console.log(data)
   }
 
   useEffect(() => {
