+++ src/pages/auth/signUp/signUpPage.module.scss
@@ -3,7 +3,6 @@
   flex-direction: column;
   align-items: center;
 
-  width: 100%;
   max-width: 378px;
   padding: 24px 24px 30px;
