+++ src/shared/router/routes.ts
@@ -1,3 +1,8 @@
 export const routes = {
   home: '/',
+  signIn: '/auth/signin',
+  legal: {
+    termsOfService: '/legal/terms-of-serivce',
+    privacyPolicy: '/legal/privacy-policy',
+  },
 }
